Rename to opencensusexporter
