+++ ext/opentelemetry-ext-opencensusexporter/MANIFEST.in
