+++ ext/opentelemetry-ext-opencensusexporter/README.rst
@@ -3,8 +3,8 @@ OpenTelemetry Collector Exporter
 
 |pypi|
 
-.. |pypi| image:: https://badge.fury.io/py/opentelemetry-ext-opencensus.svg
-   :target: https://pypi.org/project/opentelemetry-ext-opencensus/
+.. |pypi| image:: https://badge.fury.io/py/opentelemetry-ext-opencensusexporter.svg
+   :target: https://pypi.org/project/opentelemetry-ext-opencensusexporter/
 
 This library allows to export data to `OpenTelemetry Collector`_ , currently using OpenCensus receiver in Collector side.
 
@@ -13,12 +13,12 @@ Installation
 
 ::
 
-     pip install opentelemetry-ext-opencensus
+     pip install opentelemetry-ext-opencensusexporter
 
 
 References
 ----------
 
-* `OpenTelemetry Collector Exporter <https://opentelemetry-python.readthedocs.io/en/latest/ext/opencensus/opencensus.html>`_
+* `OpenTelemetry Collector Exporter <https://opentelemetry-python.readthedocs.io/en/latest/ext/opencensusexporter/opencensusexporter.html>`_
 * `OpenTelemetry Collector <https://github.com/open-telemetry/opentelemetry-collector/>`_
 * `OpenTelemetry <https://opentelemetry.io/>`_
