+++ ext/opentelemetry-ext-opencensusexporter/setup.cfg
@@ -13,13 +13,13 @@
 # limitations under the License.
 #
 [metadata]
-name = opentelemetry-ext-opencensus
+name = opentelemetry-ext-opencensusexporter
 description = OpenTelemetry Collector Exporter
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 author = OpenTelemetry Authors
 author_email = cncf-opentelemetry-contributors@lists.cncf.io
-url = https://github.com/open-telemetry/opentelemetry-python/tree/master/ext/opentelemetry-ext-opencensus
+url = https://github.com/open-telemetry/opentelemetry-python/tree/master/ext/opentelemetry-ext-opencensusexporter
 platforms = any
 license = Apache-2.0
 classifiers =
@@ -41,7 +41,7 @@ package_dir=
 packages=find_namespace:
 install_requires =
     grpcio >= 1.0.0, < 2.0.0
-    opencensus-proto >= 0.1.0, < 1.0.0
+    opencensusexporter-proto >= 0.1.0, < 1.0.0
     opentelemetry-api == 0.8.dev0
     opentelemetry-sdk == 0.8.dev0
     protobuf >= 3.8.0
